refactor: refactoring Anotacoes
