+++ src/components/Notes/StepsLoadingSpinner.tsx
@@ -1,0 +1,10 @@
+function StepsLoadingSpinner() {
+    return (
+        <svg className="animate-spin h-4 w-4 text-primary-500" viewBox="0 0 24 24">
+            <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4" fill="none" />
+            <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8v4a4 4 0 00-4 4H4z" />
+        </svg>
+    );
+}
+
+export default StepsLoadingSpinner;
